refactor: move typer cli app to separated cli.py
